requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

# other requirements
aicsimageio==4.11.0
aiobotocore==2.4.2
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
alembic==1.11.1
anndata==0.9.1
antlr4-python3-runtime==4.9.3
anyio==3.7.1
arrow==1.2.3
asciitree==0.3.3
async-timeout==4.0.2
attrs==23.1.0
autograd==1.6.2
autopage==0.5.1
backoff==2.2.1
beautifulsoup4==4.12.2
bfio==2.3.0
blessed==1.20.0
blinker==1.6.2
botocore==1.27.59
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
cliff==4.3.0
cloudpickle==2.2.1
cmaes==0.9.1
cmake==3.26.4
cmd2==2.4.3
colorlog==6.7.0
contourpy==1.1.0
croniter==1.4.1
cycler==0.11.0
cyto-dl==0.0.0
dask==2023.5.0
databricks-cli==0.17.7
dateutils==0.6.12
deepdiff==6.3.1
distributed==2023.5.0
docker==6.1.3
e3nn==0.5.1
einops==0.6.1
elementpath==4.1.4
entrypoints==0.4
escnn==1.0.11
exceptiongroup==1.1.2
fastapi==0.100.0
fasteners==0.18
filelock==3.12.2
fire==0.5.0
Flask==2.3.2
fonttools==4.41.0
frozenlist==1.4.0
fsspec==2023.3.0
future==0.18.3
gitdb==4.0.10
gitpython==3.1.32
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
h5py==3.9.0
hydra-colorlog==1.2.0
hydra-core==1.3.2
hydra-optuna-sweeper==1.2.0
idna==3.4
imagecodecs==2023.3.16
imageio==2.31.1
importlib-metadata==6.8.0
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
kiwisolver==1.4.4
lazy-loader==0.3
lie-learn==0.0.1.post1
lightning==2.0.5
lightning-cloud==0.5.37
lightning-utilities==0.9.0
lit==16.0.6
locket==1.0.0
lxml==4.9.3
Mako==1.2.4
markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
mlflow==2.5.0
monai-weekly==1.3.dev2329
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
natsort==8.4.0
networkx==3.1
numcodecs==0.11.0
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
ome-types==0.4.0
ome-zarr==0.8.0
omegaconf==2.3.0
opt-einsum==3.3.0
opt-einsum-fx==0.1.4
optuna==2.10.1
ordered-set==4.1.0
ostat==0.2
packaging==23.1
pandas==1.5.3
partd==1.4.0
pbr==5.11.1
pillow==10.0.0
PrettyTable==3.8.0
protobuf==3.20.1
psutil==5.9.5
py3nj==0.1.2
pyarrow==10.0.1
pydantic==1.10.11
pygments==2.15.1
pyjwt==2.7.0
pymanopt==2.1.1
pyparsing==3.0.9
pyperclip==1.8.2
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.5
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
querystring-parser==1.2.4
readchar==4.0.5
requests==2.31.0
resource-backed-dask-array==0.1.0
rich==13.4.2
s3fs==2023.3.0
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
setuptools==68.0.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlalchemy==2.0.19
sqlparse==0.4.4
starlette==0.27.0
starsessions==1.3.0
stevedore==5.1.0
sympy==1.12
tabulate==0.9.0
tblib==2.0.0
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.7.10
toolz==0.12.0
torch==2.0.1
torchmetrics==1.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing-extensions==4.7.1
universal-pathlib==0.0.23
urllib3==1.26.16
uvicorn==0.23.1
wcwidth==0.2.6
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.3.6
wheel==0.40.0
wrapt==1.15.0
xarray==2023.1.0
xmlschema==2.3.1
xsdata==23.6
yarl==1.9.2
zarr==2.15.0
zict==3.0.0
zipp==3.16.2
torchvision
torch-scatter
torch-geometric
pyntcloud
plotly
point_cloud_utils
ipykernel
